module5-task1: декомпозиция main презентера;
